Deduplicated filter code
